backend/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14
pandas>=2.0
numpy>=1.24
ReportLab>=4.0
gunicorn>=21.0
psycopg2-binary>=2.9